Leave conf.env_matrix in the form it is in the config file
